pyproject.toml
[tool.poetry]
name = "warehouse-manager"
version = "0.1.0"
description = ""
authors = ["Nick <nickphilomath@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
dj-database-url = "^2.1.0"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
